server/requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiofiles>=23.2.1
python-multipart
pypdf>=3.17.0
tiktoken>=0.5.0
faiss-cpu>=1.7.4
numpy>=1.24.0
fpdf2>=2.7.0
